refactor:added userDetailSection subcomponent in common folder
